chore: moved swaps endpoint to v2
